refactor: Calculating移動
